+++ frontend/src/app/components/crudifier/crudifier-frontend/crudifier-frontend.component.html
@@ -1,147 +1,51 @@
 
-<div class="form-wrapper">
+<mat-form-field class="form-field" appearance="outline">
 
-  <mat-form-field class="form-field" appearance="outline">
+  <mat-label>Select template</mat-label>
+  <mat-select
+    [(ngModel)]="template"
+    (selectionChange)="templateChanged()"
+    matTooltip="Template to generate frontend from">
+    <mat-option
+      *ngFor="let idx of templates"
+      [value]="idx">{{idx}}</mat-option>
+  </mat-select>
 
-    <mat-label>Select template</mat-label>
-    <mat-select
-      [(ngModel)]="template"
-      (selectionChange)="templateChanged()"
-      matTooltip="Template to generate frontend from">
-      <mat-option
-        *ngFor="let idx of templates"
-        [value]="idx">{{idx}}</mat-option>
-    </mat-select>
+</mat-form-field>
 
-  </mat-form-field>
+<mat-form-field class="form-field" appearance="outline">
 
-  <mat-form-field class="form-field" appearance="outline">
+  <mat-label>Name</mat-label>
+  <input
+    matInput
+    type="text"
+    matTooltip="Name of your app"
+    autocomplete="off"
+    [(ngModel)]="name">
 
-    <mat-label>Name</mat-label>
-    <input
-      matInput
-      type="text"
-      matTooltip="Name of your app"
-      autocomplete="off"
-      [(ngModel)]="name">
+</mat-form-field>
 
-  </mat-form-field>
+<mat-form-field class="form-field form-field-double" appearance="outline">
 
-  <mat-form-field class="form-field-double" appearance="outline">
+  <mat-label>Copyright notice</mat-label>
+  <input
+    matInput
+    type="text"
+    autocomplete="off"
+    matTooltip="Copyright notice to put in source files"
+    [(ngModel)]="copyright">
 
-    <mat-label>Docker image name</mat-label>
-    <input
-      matInput
-      type="text"
-      matTooltip="Docker image name, complete with organisation name, as you intend to upload to Docker HUB"
-      autocomplete="off"
-      [(ngModel)]="dockerImage">
+</mat-form-field>
 
-  </mat-form-field>
-
-  <mat-form-field class="form-field-double" appearance="outline">
-
-    <mat-label>Deployment URL</mat-label>
-    <input
-      matInput
-      type="text"
-      matTooltip="Frontend URL you intend to resolve application to using a DNS A record later"
-      autocomplete="off"
-      [(ngModel)]="deploymentUrl">
-
-  </mat-form-field>
-
-  <mat-form-field class="form-field form-field-double" appearance="outline">
-
-    <mat-label>Copyright notice</mat-label>
-    <input
-      matInput
-      type="text"
-      autocomplete="off"
-      matTooltip="Copyright notice to put in source files"
-      [(ngModel)]="copyright">
-
-  </mat-form-field>
+<div class="button-strip">
 
   <button
     mat-raised-button
     color="primary"
-    [disabled]="!template || name === '' || selectedEndpoints() === 0"
-    [matTooltip]="getGenerateTooltip()"
+    class="generate-button"
+    [disabled]="!template || name === ''"
     (click)="generate()">
     Generate
   </button>
 
-  <div class="module-declarations" *ngIf="template && endpoints">
-
-    <strong>Modules</strong>
-
-    <mat-chip-list>
-      <mat-chip
-        [class]="moduleSelected(idx) ? 'selected' : 'not-selected'"
-        *ngFor="let idx of modules"
-        (click)="moduleClicked(idx)">
-        {{idx}}
-        <mat-icon
-          *ngIf="moduleSelected(idx)"
-          matChipRemove>check</mat-icon>
-      </mat-chip>
-    </mat-chip-list>
-
-    <strong *ngIf="template && getComponents().length > 0">Components</strong>
-
-    <mat-chip-list *ngIf="endpoints">
-      <mat-chip
-        [class]="componentSelected(idx) ? 'selected' : 'not-selected'"
-        *ngFor="let idx of getComponents()"
-        (click)="componentClicked(idx)">
-        {{idx}}
-        <mat-icon
-          *ngIf="componentSelected(idx)"
-          matChipRemove>check</mat-icon>
-      </mat-chip>
-    </mat-chip-list>
-
-  </div>
-
-  <div
-    *ngIf="documentation && showDocumentation"
-    class="documentation"
-    [innerHtml]="documentation | marked"></div>
-
 </div>
-
-<table
-  *ngIf="endpoints && template"
-  mat-table
-  [dataSource]="endpoints"
-  class="endpoints-items clickable">
-
-  <ng-container matColumnDef="path">
-    <td
-      mat-cell
-      *matCellDef="let el">{{el.path}}</td>
-  </ng-container>
-
-  <ng-container matColumnDef="verb">
-    <td
-      mat-cell
-      class="mini-col"
-      *matCellDef="let el">{{el.verb}}</td>
-  </ng-container>
-
-  <ng-container matColumnDef="selected">
-    <td
-      mat-cell
-      class="mini-col"
-      *matCellDef="let el">
-      <mat-checkbox matTooltip="Include endpoint in application" [(ngModel)]="el.selected"></mat-checkbox>
-    </td>
-  </ng-container>
-
-  <tr
-    mat-row
-    (click)="el.selected = !el.selected"
-    *matRowDef="let el; columns: displayedColumns;"></tr>
-
-</table>
